Move component NotAvillable from pages folder to components folder
